changed conver num to base to take unsigned long, fix truncation

+++ main.h
@@ -98,6 +98,6 @@ strout_t convert_undefined(const char *, unsigned int);
 /** Helper functions */
 void reverse_string_helper(char *buffer, int len);
 strout_t handle_zero_case(void);
-strout_t convert_num_to_base(unsigned int n, char specifier);
+strout_t convert_num_to_base(unsigned long n, char specifier);
 
 #endif
